Changed INT_MAX to INT8_MAX for ubuntu compatability

+++ Assignment 2/Encapsulated ver./Cross.h
@@ -9,8 +9,8 @@ private: //additional attributes only unique to Cross
     int x_ord[12];
     int y_ord[12];
     int area = 0;
-    int x_min = INT_MAX;
-    int y_min = INT_MAX;
+    int x_min = INT8_MAX;
+    int y_min = INT8_MAX;
     int x_max = 0;
     int y_max = 0;
